chore: move ReplicationHandle to separate file
