+++ openraft/src/core/raft_core.rs
@@ -91,9 +91,9 @@ use crate::raft_state::RuntimeStats;
 use crate::raft_state::io_state::io_id::IOId;
 use crate::raft_state::io_state::log_io_id::LogIOId;
 use crate::replication::ReplicationCore;
-use crate::replication::ReplicationHandle;
 use crate::replication::ReplicationSessionId;
 use crate::replication::replication_context::ReplicationContext;
+use crate::replication::replication_handle::ReplicationHandle;
 use crate::replication::request::Replicate;
 use crate::replication::snapshot_transmitter::SnapshotTransmitter;
 use crate::runtime::RaftRuntime;
